refactor: Improve project class
Use a when expression instead of an if-else statement in the srcDir property. This will make the code more concise and easier to read.

Add proper documentation for each function explaining what it does and how it can be used.

Use a data class instead of a regular class for the Project class. This will provide default implementations of hashCode, equals, and toString methods that are based on the class properties.