refactor: Improve project adapter class
Instead of using lateinit to set the listener, we can pass it as a parameter to the constructor of the adapter. This way, we can ensure that the listener is always set before the adapter is used.

Instead of directly accessing and modifying the project list in the adapter, we can use a backing field to encapsulate it. This way, we can ensure that any changes to the list are properly handled and notified to the adapter.