refactor: Improve FileIndex class
Use File.separator instead of File.pathSeparator when constructing the path to the cache directory.

Use a single list comprehension to convert the list of files to a list of file paths.

Use a try-with-resources block when creating and writing to the file to ensure that the file is properly closed after use.

Use a more specific type for the list of files when deserializing the JSON data.

Use a more descriptive name for the variable holding the list of file paths.