feat: move the project loading process to the viewmodel
Removed the setOnClickListeners(), setUpProjectList(), observeViewModelProjects() and showSnackbarIfNoProjects() methods to reduce the complexity of the onViewCreated() method
The findNavController() request has been moved to a separate navigateToNewProjectFragment() and navigateToEditorFragment() method to reduce repetitive code.

To improve the readability of the code, the if blocks have also been moved to separate methods.