feat(app): Optimize `App.kt` class Create a separate function called `extractAsset` to avoid repetitive code in the `extractFiles` function.
Remove unnecessary `withContext` calls in the `disableModules` and `loadTextmateTheme` functions since they are already being called from a coroutine with the `Dispatchers.IO` context.

Create a separate function called `applyThemeBasedOnConfiguration` to apply the theme based on the configuration, which is used in both `loadTextmateTheme` and `onConfigurationChanged`. This will avoid code duplication.

Signed-off-by: TheRemakerMan <farzasz7@gmail.com>